[Fix] Crashes in plugin bridge due to re-used chunk file names
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@5693 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ pluginBridge/BridgeWrapper.h
@@ -63,6 +63,7 @@ protected:
 	Event sigAutomation;
 	MappedMemory oldProcessMem;
 
+	// Helper struct for keeping track of auxiliary shared memory
 	struct AuxMem
 	{
 		LONG used;
@@ -132,7 +133,7 @@ protected:
 	void DispatchToHost(DispatchMsg *msg);
 	bool SendToBridge(BridgeMessage &sendMsg);
 	void SendAutomationQueue();
-	AuxMem *GetAuxMemory(size_t size);
+	AuxMem *GetAuxMemory(uint32 size);
 
 	static VstIntPtr VSTCALLBACK DispatchToPlugin(AEffect *effect, VstInt32 opcode, VstInt32 index, VstIntPtr value, void *ptr, float opt);
 	static void VSTCALLBACK SetParameter(AEffect *effect, VstInt32 index, float parameter);
